Getting ready for final 1.7.10 RC build

+++ src/main/java/net/machinemuse/utils/MPSTeleporter.java
@@ -105,10 +105,10 @@ public class MPSTeleporter extends Teleporter {
             long j = par1 - 600L;
             while (iterator.hasNext()) {
                 Long olong = (Long) iterator.next();
-                PortalPosition portalposition = (PortalPosition) this.destinationCoordinateCache.getValueByKey(olong);
+                PortalPosition portalposition = (PortalPosition) this.destinationCoordinateCache.getValueByKey(olong.longValue());
                 if ((portalposition == null) || (portalposition.lastUpdateTime < j)) {
                     iterator.remove();
-                    this.destinationCoordinateCache.remove(olong);
+                    this.destinationCoordinateCache.remove(olong.longValue());
                 }
             }
         }
